requirements.txt
streamlit
tensorflow
keras
pandas
numpy
matplotlib
opencv-python
fpdf
psycopg2-binary
Pillow
